Improve implementation of ConfigurationProviders...
...and scopes.
